api/requirements.txt
psycopg2-binary
fastapi[all]
uvicorn
pytest
sqlalchemy 
bcrypt 
pydantic 
python-jose